+++ ll/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `defmt::Display` for `fdm::W5500`, `vdm::W5500`, and `vdm::Error`.
 - Added `Protocol::from_raw` to convert from a raw value in `const` contexts.
 - Added `OperationMode::from_raw` to convert from a raw value in `const` contexts.
+- Added `TryFrom<u8> for SocketCommand`.
 
 ### Changed
 - Changed all field setters for registers to methods that return a modified structure.